tech(api): rename scoring-v3.js into scoring-v3-old.js
