+++ api/src/certification/evaluation/domain/services/scoring/scoring-v3-old.js
